remake main.dart and add MyHomePage.dart file
